memfs: When using O_CREATE, the file is allowed to exist, unless O_EXCL
OpenFile currently gives an error if O_CREATE is passed and the file
already exists. This is not how O_CREATE is meant to work. O_CREATE
requests the file to be created if it does not exist. However, if O_EXCL
is also set, then the file must not exist.

Implement and test this behavior.